Refactor code, try to add proper ui
