+++ src/main.rs
@@ -9,6 +9,7 @@ mod board;
 mod engine;
 mod cam;
 mod structs;
+mod ui;
 
 use cam::Cam;
 use structs::{CellProp, GameWindow};
@@ -23,6 +24,7 @@ use opengl_graphics::glyph_cache::GlyphCache;
 
 use engine::Engine;
 use board::{Board, CellDesc};
+use ui::{GameBoard};
 use std::time::{Instant, Duration};
 use std::cell::Cell;
 
@@ -50,7 +52,7 @@ struct Resources {
 
 struct Game<'a,> {
 
-    window: GameWindow,
+    //window: GameWindow,
     cell: CellProp,
 
     show_grid: bool,
@@ -61,6 +63,8 @@ struct Game<'a,> {
     cur_state: State,
 
     resources: Resources,
+
+    ui_manager: ui::UI,
 
 }
 
@@ -77,9 +81,9 @@ impl<'a> Game<'a> {
             .build()
             .unwrap();
 
-        Game {
+        let mut game = Game {
 
-            window: GameWindow::new(width, height, window),
+            //window: GameWindow::new(width, height, window),
 
             cell: CellProp::new(10.0, 10.0),
 
@@ -89,7 +93,7 @@ impl<'a> Game<'a> {
             // enable/disable rendering
             render: true,
 
-            engine: Engine::new(Some(100), Some(100)),
+            engine: Engine::new(Some(200), Some(200)),
 
             cam: Cam::new(0.0, 0.0),
 
@@ -102,164 +106,20 @@ impl<'a> Game<'a> {
                     join("Roboto-Regular.ttf")).unwrap(),
             },
 
-        }
+            ui_manager: ui::new(GameWindow::new(width, height, window)),
+
+        };
+//
+//        let game_board_window = ui::new_board_window(GameWindow::new(width, height, window),
+//                                                     Engine::new(Some(200), Some(200)));
+//
+//        game.ui_manager.push(Box::new(game_board_window));
+
+        game
     }
 
     fn event_dispatcher(&mut self) {
-        let mut last_iter_time = Instant::now();
-        let mut last_pos: Option<[f64; 2]> = None;
-
-        let mut gl = GlGraphics::new(OPENGL);
-
-        loop {
-
-            let event = { self.window.get_window().borrow_mut().next() };
-
-            match event {
-                Some(e) => {
-                    match e {
-                        Event::Render(args) => {
-                            gl.draw(args.viewport(), |c, g| self.paint(c, g));
-                        }
-
-                        Event::Update(_) => {
-                            if self.cur_state == State::Working || self.cur_state == State::StepByStep {
-                                if !self.render ||
-                                    Instant::now() - last_iter_time >= Duration::from_millis(3) ||
-                                    self.cur_state == State::StepByStep {
-
-                                    self.engine.iterations(1);
-                                    last_iter_time = Instant::now();
-
-                                    if self.cur_state == State::StepByStep {
-                                        self.cur_state = State::Paused;
-                                    }
-
-                                }
-                            }
-                        }
-
-                        Event::Input(Input::Press(Button::Keyboard(Key::P))) => {
-                            // pause/unpause
-                            if self.cur_state == State::Working {
-                                self.cur_state = State::Paused;
-                                // always enable rendering in pause mode
-                                self.render = true;
-                            } else {
-                                self.cur_state = State::Working;
-                            }
-                        }
-
-                        Event::Input(Input::Press(Button::Keyboard(Key::S))) => {
-                            // step by step mode
-                            if self.cur_state == State::Working || self.cur_state == State::Paused {
-                                self.cur_state = State::StepByStep;
-                                // always enable rendering in step by step mode
-                                self.render = true;
-                            }
-                        }
-
-                        Event::Input(Input::Press(Button::Keyboard(Key::G))) => {
-                            // show/hide grid
-                            self.show_grid = !self.show_grid;
-                        }
-
-
-                        Event::Input(Input::Press(Button::Mouse(MouseButton::Left))) => {
-                            self.cur_state = State::Draw;
-                        }
-
-                        Event::Input(Input::Release(Button::Mouse(MouseButton::Left))) => {
-                            if last_pos.is_some() {
-                                let pos = last_pos.unwrap();
-                                self.born_or_kill(true, pos[0], pos[1]);
-
-                                self.cur_state = State::Paused;
-                            }
-                        }
-
-                        Event::Input(Input::Move(Motion::MouseCursor(x, y))) => {
-                            if self.cur_state == State::Draw {
-                                self.born_or_kill(false, x, y);
-                            }
-                            last_pos = Some([x, y]);
-                        }
-
-                        // movements control ->
-                        Event::Input(Input::Press(Button::Keyboard(Key::Right))) => {
-                            self.cam.move_right();
-                        }
-
-                        Event::Input(Input::Release(Button::Keyboard(Key::Right))) => {
-                            self.cam.reset_move_step();
-                        }
-
-                        Event::Input(Input::Press(Button::Keyboard(Key::Left))) => {
-                            self.cam.move_left();
-                        }
-
-                        Event::Input(Input::Release(Button::Keyboard(Key::Left))) => {
-                            self.cam.reset_move_step();
-                        }
-
-                        Event::Input(Input::Press(Button::Keyboard(Key::Up))) => {
-                            self.cam.move_up();
-                        }
-
-                        Event::Input(Input::Release(Button::Keyboard(Key::Up))) => {
-                            self.cam.reset_move_step();
-                        }
-
-                        Event::Input(Input::Press(Button::Keyboard(Key::Down))) => {
-                            self.cam.move_down();;
-                        }
-
-                        Event::Input(Input::Release(Button::Keyboard(Key::Down))) => {
-                            self.cam.reset_move_step();
-                        }
-                        // movements control <-
-
-                        // zoom out ->
-                        Event::Input(Input::Press(Button::Keyboard(Key::NumPadMinus))) => {
-                            self.cam.zoom_out();
-                        }
-
-                        Event::Input(Input::Press(Button::Keyboard(Key::Minus))) => {
-                            self.cam.zoom_out();
-                        }
-                        // zoom out <-
-
-                        // zoom in ->
-                        Event::Input(Input::Press(Button::Keyboard(Key::NumPadPlus))) => {
-                            self.cam.zoom_in();
-                        }
-
-                        // use "Equals" instead of "Plus" to avoid holding shift key requirement
-                        Event::Input(Input::Press(Button::Keyboard(Key::Equals))) => {
-                            self.cam.zoom_in();
-                        }
-                        // zoom in <-
-
-                        Event::Input(Input::Press(Button::Keyboard(Key::R))) => {
-                            // If in pause mode - fill board with a random pattern
-                            if self.cur_state == State::Paused {
-                                let board = self.engine.create_random(0.3);
-                                self.engine.set_board(board);
-                            }
-                            else {
-                                // otherwise enable/disable rendering
-                                self.render = !self.render;
-                            }
-
-                        }
-
-                        _ => {}
-                    }
-                }
-
-                None => break
-            }
-        }
+        self.ui_manager.event_dispatcher();
     }
 
     fn born_or_kill(&mut self, kill_alive: bool, x: f64, y: f64) {
