+++ src/ui/window/mod.rs
@@ -1,0 +1,20 @@
+pub mod confirm;
+pub mod board;
+
+pub trait WindowBase {
+
+    fn paint(&self);
+
+}
+
+pub trait ActiveWindow: WindowBase {
+
+    fn event_dispatcher(&self);
+
+}
+
+
+pub trait InactiveWindow: WindowBase {
+
+
+}
